+++ lib/vrl/compiler/src/expression/object.rs
@@ -1,7 +1,7 @@
 use crate::expression::{Expr, Resolved};
 use crate::{Context, Expression, State, TypeDef, Value};
 use std::fmt;
-use vrl_structures::Map;
+use vrl_structures::map::Map;
 
 #[derive(Debug, Clone, PartialEq)]
 pub struct Object {
