+++ lib/vrl/compiler/src/type_def.rs
@@ -1,7 +1,7 @@
 use crate::value::Kind;
 use crate::{map, path, Path};
 use std::{collections::BTreeSet, ops::Sub};
-use vrl_structures::Map;
+use vrl_structures::map::Map;
 
 /// Properties for a given expression that express the expected outcome of the
 /// expression.
