+++ lib/vrl/stdlib/src/compact.rs
@@ -1,6 +1,6 @@
 use crate::util;
 use vrl::prelude::*;
-use vrl_structures::Map;
+use vrl_structures::map::Map;
 
 #[derive(Clone, Copy, Debug)]
 pub struct Compact;
