+++ lib/vrl/stdlib/src/log_util.rs
@@ -2,7 +2,7 @@ use chrono::prelude::*;
 use lazy_static::lazy_static;
 use regex::{Captures, Regex};
 use vrl::prelude::*;
-use vrl_structures::Map;
+use vrl_structures::map::Map;
 
 lazy_static! {
     // Information about the common log format taken from the
