+++ lib/vrl/stdlib/src/parse_aws_vpc_flow_log.rs
@@ -1,5 +1,5 @@
 use vrl::prelude::*;
-use vrl_structures::Map;
+use vrl_structures::map::Map;
 
 #[derive(Clone, Copy, Debug)]
 pub struct ParseAwsVpcFlowLog;
