+++ lib/vrl/stdlib/src/parse_common_log.rs
@@ -1,6 +1,6 @@
 use crate::log_util;
 use vrl::prelude::*;
-use vrl_structures::Map;
+use vrl_structures::map::Map;
 
 #[derive(Clone, Copy, Debug)]
 pub struct ParseCommonLog;
