+++ lib/vrl/stdlib/src/parse_glog.rs
@@ -2,7 +2,7 @@ use chrono::{offset::TimeZone, Utc};
 use lazy_static::lazy_static;
 use regex::Regex;
 use vrl::prelude::*;
-use vrl_structures::Map;
+use vrl_structures::map::Map;
 
 lazy_static! {
     static ref REGEX_GLOG: Regex = Regex::new(
