+++ lib/vrl/stdlib/src/parse_url.rs
@@ -1,7 +1,7 @@
 use std::iter::FromIterator;
 use url::Url;
 use vrl::prelude::*;
-use vrl_structures::Map;
+use vrl_structures::map::Map;
 
 #[derive(Clone, Copy, Debug)]
 pub struct ParseUrl;
